+++ src/main/java/com/thf/dao/FunctionDAO.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface FunctionDAO {
     int insert(Function function);
     int update(Function function);
-    Function searchById(int fId);
-    int delete(int id);
-    int deleteAll(int mid);
-    List<Function> searchAllFunction(int mid);
+    Function searchById(long fId);
+    int delete(long id);
+    int deleteAll(long mid);
+    List<Function> searchAllFunction(long mid);
 }
