+++ src/main/java/com/thf/dao/ProjectDAO.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface ProjectDAO {
      int insertProject(Project project);
      int updateProject(Project project);
-     Project searchById(int id);
+     Project searchById(long id);
      List<Project> searchKey(@Param("key") String key);
      List<Project> getAllProject(Project project);
 }
