+++ src/main/java/com/thf/service/FunctionService.java
@@ -4,8 +4,8 @@ import com.thf.common.oo.ResultVO;
 import com.thf.entity.Function;
 
 public interface FunctionService {
-    ResultVO createFunc(String token,int mid, Function function);
+    ResultVO createFunc(String token,long mid, Function function);
     ResultVO updateFunc(String token,Function function);
-    ResultVO deleteFunc(String token,int[] fidlist);
-    ResultVO searchFunc(int fid);
+    ResultVO deleteFunc(String token,long[] fidlist);
+    ResultVO searchFunc(long fid);
 }
