+++ src/main/java/com/thf/service/ProjectService.java
@@ -6,7 +6,7 @@ import com.thf.entity.Project;
 public interface ProjectService {
     ResultVO createProject(String token,Project project);
     ResultVO updateProject(String token,Project project);
-    ResultVO searchById(int id);
+    ResultVO searchById(long id);
     ResultVO searchKey(String key);
     ResultVO deleteProject(String token,Project project);
     ResultVO getAllProject(String token);
